change datatype of total_competitors from BIGINT to INT

+++ apps/openchallenges/challenge-service/src/main/resources/db/migration/V1.0.0__create_tables.sql
@@ -33,7 +33,7 @@ CREATE TABLE challenge (
   operation_id BIGINT,
   created_at TIMESTAMP WITH TIME ZONE NOT NULL,
   updated_at TIMESTAMP WITH TIME ZONE NOT NULL,
-  total_competitors BIGINT,
+  total_competitors INT,
   total_teams BIGINT,
   total_submissions BIGINT,
   CONSTRAINT fk_platform FOREIGN KEY (platform_id) REFERENCES challenge_platform(id),
